+++ team2/pom.xml
@@ -103,10 +103,10 @@
   		    <version>1.15.3</version>
   		</dependency>
   		<dependency>
-  			<groupId>com.example</groupId>
-  			<artifactId>board</artifactId>
-  			<version>0.0.1-SNAPSHOT</version>
-  		</dependency>
+			  <groupId>com.example</groupId>
+			  <artifactId>board</artifactId>
+			  <version>0.0.1-SNAPSHOT</version>
+		  </dependency>
 	</dependencies>
 
 	<build>
