+++ team2/src/main/resources/application.properties
@@ -14,7 +14,7 @@ spring.servlet.multipart.enabled=true
 spring.servlet.multipart.max-file-size=50MB
 spring.servlet.multipart.max-request-size=50MB
 
-spring.security.user.password.encoder=bcrypt
+#spring.security.user.password.encoder=bcrypt
 
 # com.example.demo \uC778\uD130\uD398\uC774\uC2A4\uAC00 \uC788\uB294 \uD328\uD0A4\uC9C0, \uC774 \uD328\uD0A4\uC9C0 \uC548\uC758 \uC778\uD130\uD398\uC774\uC2A4\uB294 @Repository, @Mapper\uAC00 \uC124\uC815\uB418\uC5B4 \uC788\uC5B4\uC57C \uD568
 mybatis.type-aliases-package=com.team2
